Remove Route.GetPath and use a factory method to export as graph instead
