+++ src/Caravel.Abstractions/IRoute.cs
@@ -4,5 +4,4 @@ namespace Caravel.Abstractions;
 public interface IRoute
 {
     ImmutableList<IEdge> Edges { get; }
-    public string[] GetPath();
 }
