+++ src/Caravel.Core/Route.cs
@@ -11,20 +11,4 @@ public class Route : IRoute
     {
         Edges = edges;
     }
-
-    public string[] GetPath()
-    {
-        var path = new List<string>();
-        foreach (var edge in Edges)
-        {
-            path.Add(
-                edge.ToString()
-                    ?? throw new InvalidOperationException(
-                        $"Could not format {edge.GetType()} to string."
-                    )
-            );
-        }
-
-        return [.. path];
-    }
 }
